TelegramChatBot/requirements.txt
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
click==8.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.10.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.14
fastjsonschema==2.17.1
filelock==3.15.4
Flask==2.2.5
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.1
jsonschema==4.17.3
jupyter_core==5.3.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matrix-client==0.4.0
mdurl==0.1.2
multidict==6.0.5
nbformat==5.7.0
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
prettytable==3.10.2
pyaes==1.6.1
pyasn1==0.6.0
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0
regex==2024.5.15
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.4
smmap==5.0.1
tabulate==0.9.0
Telethon==1.36.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.19.1
traitlets==5.9.0
transformers==4.42.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.19
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
yarl==1.9.4